GMM Reduction by Williams and Maybeck (2003)

Given a GMM model with \(N\) components, where \(N\) is usually large, this function tries to find an approximation with \(M \in [1,N)\) components by iteratively merging two nearest components with respect to \(L_2\) dissimilarity measure. Merge is done via moment-preserving merge.

gmr2003W(gmmobj, M = 2)

Arguments

gmmobj
gmm-like object of \(N\) components.

M
the number of components for the reduced model.

Value

gmm-like object of \(M\) components.

References

Williams J, Maybeck P (2003). “Cost-Function-Based Gaussian Mixture Reduction for Target Tracking.” Sixth International Conference of Information Fusion, 2003. Proceedings of the, 2, 1047--1054.

Examples

# \donttest{
# -------------------------------------------------------------
#              Reduction of Mixture for SMILEY Data
#
# From multiple SMILEY data, we fit 10-component GMM for each 
# data and average of models is taken to give a GMM fit with 
# large number of redundant components.
# -------------------------------------------------------------
# Generate 20 datasets with noise and fix GMM 
list_gmm  = list()
for (i in 1:20){
  data_i = gensmiley(sd=0.25)$data
  list_gmm[[i]] = gmm(data_i, k=10)
}

# Find the average of models
gcenter = gcsum(list_gmm)

# Do Reduction for M=1,3,9
gr1 = gmr2003W(gcenter, M=1)
gr3 = gmr2003W(gcenter, M=3)
gr9 = gmr2003W(gcenter, M=9)

# Visualize using 'ggplot2'
if (requireNamespace("ggplot2", quietly = TRUE)){
  # prepare grid and density evaluation
  pgrid = as.matrix(expand.grid(x=seq(from=-1.5,to=1.5,length.out=300),
                    y=seq(from=-1.5,to=1.5,length.out=300)))
  prob1 = gmmdensity(pgrid, gr1)
  prob3 = gmmdensity(pgrid, gr3)
  prob9 = gmmdensity(pgrid, gr9)
  
  # wrap as a single dataframe
  obj1 = rbind(cbind(pgrid, prob1), cbind(pgrid, prob3), cbind(pgrid, prob9))
  obj2 = as.factor(rep(c(1,3,9),each=90000))
  odf  = data.frame(x=obj1[,1], y=obj1[,2], density=obj1[,3], class=obj2)
  levels(odf$class) = c("M=1","M=3","M=9")
  
  # visualize
  ggplot(odf, aes(x=x,y=y,z=density)) + 
    facet_grid(. ~ class) + 
    geom_raster(aes(fill=density)) + 
    geom_contour(colour="white") +
    scale_fill_viridis_c() +
    scale_x_continuous(expand = c(0, 0)) +
    scale_y_continuous(expand = c(0, 0)) +
    coord_fixed(xlim=c(-1.5,1.5), ylim=c(-1.5,1.5)) + 
    ggtitle("Reduction of 200-Component GMM via Williams and Maybeck (2003).")
}

#> Error in ggplot(odf, aes(x = x, y = y, z = density)): could not find function "ggplot"

# }